# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Automation Studio API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /webhooks/{token}:
    post:
      operationId: triggerWebhook
      tags: [workflows]
      summary: Trigger a workflow via its unique webhook URL (public, no auth)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Execution result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Execution"
        "404":
          description: Webhook not found
...
components:
  schemas:
...
    Execution:
      type: object
      properties:
        id:
          type: integer
        workflowId:
          type: integer
        workflowName:
          type: string
        status:
          type: string
          enum: [success, error, running]
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
          nullable: true
        durationMs:
          type: integer
          nullable: true
        error:
          type: string
          nullable: true
        nodeResults:
          type: array
          items:
            $ref: "#/components/schemas/NodeResult"
        inputData:
          type: object
        outputData:
          type: object
          nullable: true
      required: [id, workflowId, workflowName, status, startedAt, nodeResults, inputData]
    NodeResult:
      type: object
      properties:
        nodeId:
          type: string
        nodeType:
          type: string
        status:
          type: string
          enum: [success, error, skipped]
        output:
          type: object
          nullable: true
        error:
          type: string
          nullable: true
        durationMs:
          type: integer
      required: [nodeId, nodeType, status, durationMs]
...